% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_art.R
\name{make_art}
\alias{make_art}
\title{Create some random scatter plots
Reminder: ggplot2 need to be installed and loaded before using this function}
\usage{
make_art(seed)
}
\arguments{
\item{seed}{numeric variable}
}
\description{
Create some random scatter plots
Reminder: ggplot2 need to be installed and loaded before using this function
}
